+++ crates/chat-cli/src/mcp_client/mod.rs
@@ -1,16 +1,4 @@
-pub mod client;
-pub mod error;
-pub mod facilitator_types;
-pub mod messenger;
 pub mod new_client;
 pub mod new_messenger;
-pub mod server;
-pub mod transport;
 
-pub use client::*;
-pub use facilitator_types::*;
-pub use messenger::*;
 pub use new_client::*;
-#[allow(unused_imports)]
-pub use server::*;
-pub use transport::*;
