+++ crates/chat-cli/src/util/process/mod.rs
@@ -1,5 +1,3 @@
-pub use sysinfo::Pid;
-
 #[cfg(target_os = "windows")]
 mod windows;
 #[cfg(target_os = "windows")]
@@ -7,5 +5,3 @@ pub use windows::*;
 
 #[cfg(not(windows))]
 mod unix;
-#[cfg(not(windows))]
-pub use unix::*;
